rename packages to registry
